Move is_* methods from data/utils to utils
