+++ cartoframes/data/__init__.py
@@ -3,13 +3,10 @@ from __future__ import absolute_import
 
 from .dataset import Dataset
 from .dataset_info import DatasetInfo
-from .utils import is_sql_query, is_geojson_file
 from .registry.strategies_registry import StrategiesRegistry
 
 __all__ = [
     'Dataset',
     'DatasetInfo',
-    'is_sql_query',
-    'is_geojson_file',
     'StrategiesRegistry'
 ]
