+++ cartoframes/data/registry/table_dataset.py
@@ -4,7 +4,7 @@ from carto.exceptions import CartoException, CartoRateLimitException
 
 from .base_dataset import BaseDataset
 from ...columns import Column, normalize_name
-from ..utils import is_table_name
+from ...utils import is_table_name
 
 
 class TableDataset(BaseDataset):
